another step towards kexi<=>koffice integration:
kword mailmerge with kexi
and a framework for accessing kexi and loading only parts, which are needed.

svn path=/trunk/koffice/; revision=220936

+++ kexi/kexidb_old/drivers/cql/driver/cqlrecord.cpp
@@ -166,7 +166,7 @@ CqlRecord::insert()
 }
 
 void
-CqlRecord::gotoRecord(unsigned int record)
+CqlRecord::gotoRecord(unsigned long record)
 {
 }
